Move NullModelMultipleChoiceCaseInsensitiveField to base/fields.py
